requirements.txt
asgiref==3.7.2
astroid==3.1.0
autopep8==2.0.4
backports.zoneinfo==0.2.1
certifi==2024.2.2
dill==0.3.8
Django==4.1.5
django-autoslug==1.9.9
django-elasticsearch-dsl==8.0
django-js-asset==2.2.0
django-mptt==0.14.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
docstring-to-markdown==0.15
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
flake8==7.0.0
gitdb==4.0.11
GitPython==3.1.41
importlib_metadata==7.1.0
isort==5.13.2
jedi==0.19.1
Markdown==3.5.2
mccabe==0.7.0
numpy==1.24.4
packaging==24.0
pandas==2.0.3
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
pytoolconfig==1.3.1
pytz==2024.1
rope==1.13.0
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
whatthepatch==1.0.5
yapf==0.40.2
zipp==3.18.1
